# internal/adapters/storage/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type Batch struct {
	ID          string       `json:"id"`
	TotalAmount int64        `json:"total_amount"`
	TotalCount  int64        `json:"total_count"`
	Status      string       `json:"status"`
	CreatedAt   sql.NullTime `json:"created_at"`
}

type Payout struct {
	ID             string         `json:"id"`
	BatchID        sql.NullString `json:"batch_id"`
	ReferenceID    string         `json:"reference_id"`
	RecipientName  string         `json:"recipient_name"`
	RecipientPhone string         `json:"recipient_phone"`
	RecipientEmail sql.NullString `json:"recipient_email"`
	RecipientTag   sql.NullString `json:"recipient_tag"`
	CountryCode    string         `json:"country_code"`
	BankCode       sql.NullString `json:"bank_code"`
	BankName       sql.NullString `json:"bank_name"`
	AccountNumber  sql.NullString `json:"account_number"`
	Amount         int64          `json:"amount"`
	Currency       string         `json:"currency"`
	Status         string         `json:"status"`
	ErrorMessage   sql.NullString `json:"error_message"`
	CreatedAt      sql.NullTime   `json:"created_at"`
	UpdatedAt      sql.NullTime   `json:"updated_at"`
}